Move generalized circle to its own module
