+++ mobius/src/geometry/mod.rs
@@ -2,6 +2,7 @@ pub mod arc_angles;
 pub mod circle;
 pub mod circular_arc;
 pub mod double_ray;
+pub mod generalized_circle;
 pub mod line;
 pub mod line_segment;
 pub mod ray;
@@ -12,6 +13,7 @@ pub use arc_angles::*;
 pub use circle::*;
 pub use circular_arc::*;
 pub use double_ray::*;
+pub use generalized_circle::*;
 pub use line::*;
 pub use line_segment::*;
 pub use ray::*;
